Update: It Works?!
Picks up C++ header files with the TEST_CASE macro in them as valid test files and stores the TEST_CASE string as the name of the test.
All doctest unit tests set to automatically pass just whilst this is all WIP.
Next step is to try and setup the namespace.
Displays the list of doctest unit tests in the test explorer finally.
Think there's some jank going on with the test container discoverers not always refreshing? (Might need to keep an eye on this...)