Update: Run Settings
Just committing so I can work on my laptop too.
Added a new settings provider class so the run settings can be used to store paths to the output files, which means the discoverer and executor can just query the settings info and get the information they need based on the test they are handling.