Update: Run Settings + Test Owner
Tried to update the test owner to use the output exe and then tried to use the project file, but it doesn't work. I think you have to update it higher up in the callstack (in DoctestTestContainerDiscoverer/DoctestTestContainer) and then pass that down. The problem with that is this is meant to pick up header files to scan for tests, if the discovery updates the source to point to the project file instead - that isn't a header file and won't provide any test cases.

Next step is seeing if you can run the tests...