Update: Settings + Executor
It's... doing something? It doesn't seem to run every test though that I'd expect based on the output. Not sure if there's just one executor per test run? I might need a list of processes to run all of the tests together. I'm selecting Run All Tests In View - so not sure if that's a special case