Update: Executor
Runs all of the relevant doctest exe files for the selected tests. DLL tests are ignored at the moment but I need to figure out a way to link the DLL with whatever exe it is used in. Next step is to try and work out how to print more information from doctest so I can easily associated doctest exe output and parse it into information for me to update test cases and test results with. Right now it just prints out each successful assertion (e.g. each successful CHECK within a TEST_CASE and not the actual TEST_CASE result itself). There is a results summary but I need something that reports on a per TEST_CASE basis.