Update: Cleanup + Executor Test Results
Test explorer now correctly updates with the doctest test results!
This code needs a lot of cleaning up and testing to make sure it maintains the same functionality.
Next step is to clean this up and then make the error message formatting easier to read.
Then workout how to handle DLLs...