Update main.cpp
Wait no, it was doing it correctly before because the constructor was taking in the extra variables and then it calls parseArgs anyway. Mixed setup by myself so I'm going with the explicit call to applyCommandLine to make it clear it will apply any extra arguments instead of just using the context constructor.