Update: DLL Tests!
DLL doctest unit tests run correctly now and report test results to the test explorer as expected!
Noticed a bug where the error message appears to be duplicated for the DLL error message and ExeUsingDLL error message - this is most likely because they are using the same output stream; need to filter it per test case rather than per exe.