Update: Optimizations + Fixes
Refining IsTestFile further so it will only add test files that actually have ' TEST_CASE(" ' in them (this is so that it won't just add EVERY single header file in a project).
Also filtering out any folders.
Fixed mutexes for files not working correctly, noticed it was hit and miss - seems to be better to just allow the code to dispose of the mutex when needed with using.