Update: Custom Options Page
Added a tools options page for this test adapter under "Test Adapter for Doctest". Currently only has one option but I wanted to commit it in a "working" state since it seems to have been a tricky thing to get up and running. The idea is that everything should query TestAdapterOptions for whatever options the code needs to check. New options should be added to ITestAdapterOptions, implemented in TestAdapterOptions and added to the relevant options dialog page class and then queried in code wherever needed.