Update: WIP Solution Directory Issues
Attempting to start fixing issues with not being able to always get the solution directory. Sometimes the log file is created under the binary folder of the VSIX instead of the solution directory that the user has opened. Going to try setting up a fully static class instead to make sure things are only initialized once - not convinced it will work as I expect it to though...
Package will automatically load once the solution opens up.