Update: Yeah...
Starting again in a way, going to try setting this up to work from discovering executables instead and reading given pdb files. Shouldn't need as much (like the test container discoverer). Done this to also try and only include MIT/open source licensed packages/references when developing. Forgot to be vigilant in the other branches and although it works there are a couple of packages that don't appear to be open source friendly.