Update: Examples + Cleanup
DoctestTestAdapterVSIX renamed to DoctestTestAdapter.VSIX for consistency, assembly name is DoctestTestAdapter now so that when the VSIX is built it is called DoctestTestAdapter.vsix instead of DoctestTestAdapterVSIX.vsix.

Examples project/solution updated. Now works relative to where the projects are placed. I've included the examples in the DoctestTestAdapter solution as well so that any discovery/execution unit tests can be run using the same examples as the Examples.sln. Also renamed the examples directory DoctestTestAdapter.Examples for naming consistency.

Added cvdump as a thirdparty binary file (with it's MIT license file too). This will be used to gather information about the discovered executables and C++ header files.