Update: Cleanup + Mocking
Added FakeItEasy as a reference for creating mock objects and whatever else is needed for unit tests - seems to be really easy to setup and get going!

Discoverer - unit test to make sure the amount of test cases and data found for the test cases is correct for the given example.