Update: Discovery
Implemented a way to get exe and any dll dependencies test cases. This uses dumpbin /dependents to get a list of dll dependents. Then it was just a case of calling GetSourceFiles for the dll file path and adding that to the list of source files.
Also, unit test for exe and dll test discovery.