Update: runsettings + godot
Implemented custom run settings for the test adapter.
Currently, this only adds CommandArguments as an extra option but this could easily be expanded upon in future!
With the new runsettings implementation, it is now possible to discover and run godot unit tests! 🎊 Just need to add a runsettings file and make sure to add a section for Doctest like so:

<Doctest>
     <CommandArguments>"--test"</CommandArguments>
 </Doctest>

--test will be passed to the godot exe when running tests from test explorer window.