Update: Executor + Cmd Arguments
Simplified and fixed a bug with getting command arguments, it wasn't taking into account that commas might already be in the test case names. So surround the test case display name with quotes before doing any other string operations/checks on it.

Custom settings now implemented. More in depth than the previous implementation. Added discovery and executor settings which the user can populate in the .runsettings file if they want to.
DiscoverySettings:
    - SearchDirectories: Only discover tests if they are within these search directories. These are relative to the solution directory.
ExecutorSettings:
    - CommandArguments: Like before, user defined command arguments to use when running the tests.
    - ExecutableOverrides: Useful to map discovered test exes to a different exe for running tests again. E.g. app.exe can use app.console.exe instead with this setting. These are relative file paths to the solution directory.

Improved batching code so that it's more scalable depending on how much bigger the command arguments are than the current limit.