[DTA-3] [TASK]
[DTA-3] [TASK] - VS Cmd Prompt Location File Path
Created a new function GetVSInstallDirectory to get the full directory of the given VS version. E.g. gets the directory for Visual Studio Community 2022. Think it would be fairly simple to expand this and provide a different version (e.g. pro instead of community) by using a different versionName.

Fixes #3