Update: Exceptions
Went through the code how I expect the flow to go from a users point of view and added exceptions where needed for critical points of information. Should be able to see errors obviously as a user when exceptions are thrown from this test adapter now - these will be visible in the test output window (and I believe the test explorer window prints single lines of text at the top which also mentions if an exception occured). These situations are unlikely and shouldn't happen but the whole point of having these here is because this will be out in the wild and I might have overlooked something!