Update: godot + Fixes
Added a new solution DoctestTestAdapter.Godot.sln which will provide developers with everything the DoctestTestAdapter.sln does, but also adds the godot engine source code as an example project and accompanying unit tests in DoctestTestAdapter.Tests.Godot. Currently this just makes sure the discovery/executor works correctly with the godot source code as well. There are some test cases missing (see the TODO comment in GodotDoctestTestDiscovererTest). Will work on this next, wanted to get this committed as this has quite a lot of changes but the test adapter works with godot; it's just missing 100 test cases. 😅

Fixed a bug with the class keyword (noticed thanks to the godot engine source code). Keywords weren't taking into account that even if the file line read was within the keyword scope, they might not necessarily be within brackets straight away - so keywords/bracket matching now checks if the first open bracket has been found.

Also pretty neat I've discovered a workaround for having '.' in the test names which seems to work okay, using "\u2024". So test case names are allowed to have '.' in them.