Update: New Keywords
Added more keywords: CustomMacroKeyword, DoctestTestCaseFixtureKeyword, DoctestTestCaseMayFailKeyword, DoctestTestCaseTemplateKeyword.
* CustomMacroKeyword - This is responsible for tracking any custom macros wrapped around doctest macros. See CUSTOM_TEST_CASE_MACRO in TestIsEvenUsingDoctestMain.h. This will automatically detect the definition of the macro and how it is mapped to the underlying doctest macro. Implemented this to be able to handle situations like "#define GETTER_TEST(m_name, m_getter, m_default_value)" from "test_fontfile.h" in the godot engine example. It works by basically reusing the existing keyword classes to implement any logic when the macro is used. So a macro that wraps around a TEST_CASE will end up calling DoctestTestCaseKeyword.
* DoctestTestCaseFixtureKeyword - Implemented to find any TEST_CASE_FIXTURE setups.
* DoctestTestCaseMayFailKeyword - Honestly didn't realise this wasn't a standard doctest keyword until after I implemented it and started working on the unit tests. 😅 Keeping it included anyway and have bundled the accompanying unit test inside of the DoctestTestAdapter.Tests.Godot project. Finds any TEST_CASE_MAY_FAIL macros.
* DoctestTestCaseTemplateKeyword - Finds any TEST_CASE_TEMPLATE macros.

Added IKeyword to be able to properly setup the custom macro keywords. The logic is quite different from the regular Keyword class. Needed to be able to have a collection of them all but not affect the current inheriting heirarchy. Macros only need to check if they end with \ to know they are on multiple lines for defining a wrapper around existing doctest macros.

Updated unit tests with new expected test numbers and lines etc.

Base test for godot unit tests in this test adapter code. Helpful to group common logic together, currently handles loading and asserting the runsettings as well as checking for any missing test cases (at the time of writing, none are missing).

Added an FAQ.md to track what might be common questions/answers for the test adapter whilst they are fresh in my mind.