Update: Bug Fix!
During testing I noticed a bug where it appeared to discover double the amount of tests when I opened a different godot solution. Turns out that because I basically copied and pasted this godot solution from my examples folder into a different folder, it still had information about the examples folder godot solution and was picking up header files from it. This happened because one of the conditions for checking search directories when getting source files didn't check to see if it was still under the solution directory. Checking if the source file contains the solution directory fixed this issue.

Added image to FAQ and section about the warning no test is available in [filename].dll.