Update DoctestExecutable.cs
Bug fix, using the same search string for start/end index and then trimming the output to get rid of any excess. This fixes an issue where it was throwing an out of range exception if there are no test suites or test cases in an application. The doctest executable will always print some kind of output, but there may still be no test suites or test cases.